WIP: Read/Write Macros split
Read/Write is now split between Scala 2&3, Scala 2 compiles as before, Scala 3 is still missing an implementation.
